Coding Challenge for Camp Gladiator
Trainer API Coding Challenge
Requirements
Create two REST API endpoints as follows:

Create a new trainer
Get details for a trainer
The Trainer model should include (at the very least) the following attributes:

{
    "id" : "trainer-id-000001",
    "email" : "trainer@campgladiator.com",
    "phone" : "5125125120",
    "first_name": "Fearless",
    "last_name": "Contender"
}

Challenge Goal
The goal of this challenge is to produce a Production ready REST API. Please don't dedicate more than 3 hours to work on this project. It's ok to not complete it in full, but please add a README with instructions to run it, and what would be your next steps if you had more time available.

Challenge Submission
Please send an email to the hiring manager including a Github link to the code you created at your earliest convenience.